[black] Add black code formatter support #86
This allows us to define a common code formatting standard and avoids
having to manually do these changes manually.
It is enabled by default and can be disabled using --skip-black.
This PR also formats the code in this repository using black.
Add isort extra configuration so that it is compatible with black.
Add convenient `openwisp-qa-format` script for code formatting in OpenWISP
projects.

Closes #86