[openwisp_utils/storage] Add a Custom Storage Class #166
Ported changes from storage.py in ansible-openwisp2 to storage.py in openwisp_utils/ .

Added documentation and written test for `CompressStaticFilesStorage` in openwisp_utils.storage
In documentation created a separate storage utilities section and in settings added `OPENWISP_STATICFILES_VERSIONED_EXCLUDE`.

Created test_storage in tests/test_project/tests , The test creates a temporary folder and 2 text files in it and then runs collectstatic in that folder and checks whether the file hashed is correct or not and later deletes that folder.

Added `django-compress-staticfiles~=1.0.1b` in install_requires in setup.py.

Updated with requested changes.

Fixes #166